requirements.txt
python-telegram-bot==20.3
Faker==18.9.0
